requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.12
Flask-PyMongo==2.3.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
limits==4.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.17.0
ordered-set==4.1.0
packaging==24.2
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.5.0
python-dotenv==1.0.1
rich==13.9.4
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
Werkzeug==3.0.1
wrapt==1.17.2
